// apps/console/src/__generated__/core/LinkedAuditsDialogFragment.graphql.ts
export type AuditState = "COMPLETED" | "IN_PROGRESS" | "NOT_STARTED" | "OUTDATED" | "REJECTED";
import { FragmentRefs } from "relay-runtime";
export type LinkedAuditsDialogFragment$data = {
  readonly audits: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly framework: {
          readonly id: string;
          readonly name: string;
        } | null | undefined;
        readonly id: string;
        readonly name: string | null | undefined;
        readonly state: AuditState;
      };
    }>;
  };
  readonly id: string;
  readonly " $fragmentType": "LinkedAuditsDialogFragment";
};
export type LinkedAuditsDialogFragment$key = {
  readonly " $data"?: LinkedAuditsDialogFragment$data;
  readonly " $fragmentSpreads": FragmentRefs<"LinkedAuditsDialogFragment">;
};